Move from report to export naming and move code to report class
