+++ csa-app/util/Shared.go
@@ -21,9 +21,12 @@ var (
 	RulesDir           = App.Flag("rules-dir", "directory where csa rules are. Rules found in this directory will be automatically imported on tool startup. This will also be the default directory for `rules` import").Default(DEFAULT_RULES_DIR).String()
 	ModelsDir          = App.Flag("models-dir", "directory where csa scoring models are. Scoring Models found in this directory will be automatically imported on tool startup. This will also be the default directory for `scoring-models` import").Default(DEFAULT_MODELS_DIR).String()
 	OutputDir          = App.Flag("output-dir", "directory path where csa results will be output").Default(DEFAULT_OUTPUT_DIR).String()
-	ReportDir          = App.Flag("report-dir", "directory path where csa Report outputs will be written").Default(DEFAULT_REPORT_DIR).String()
-	HtmlReportFileName = App.Flag("html-report-file-name", "name of the HTML report file").String()
-	CsvReportFileName  = App.Flag("csv-report-file-name", "name of the CSV report file").String()
+	
+	//Export Results Configurations
+	ExportFormats      = App.Flag("export", "list of expected formats divided by commas that will be used to export findings, ex: csv or csv,html").String()
+	ExportDir          = App.Flag("export-dir", "directory path where csa finding exports will be written").Default(DEFAULT_EXPORT_DIR).String()
+	ExportFileName     = App.Flag("export-file-name", "base name of the \"export\" file, ex: \"csa-export\". Proper extensions will be appended based on \"--export\" command formats requested.").Default(DEFAULT_EXPORT_FILE_NAME).String()
+
 	ExcludedDirsRegEx  = App.Flag(EXCLUDED_DIRS_FLAG, "regex pattern of directories not to be included in analysis").Default("^([.].*|target|bin|test|node_modules|eclipse|out|vendors|obj)$").String()
 	DB                 = App.Flag("db", "which database engine to use (sqlite|postgres)").Default(SQLITE).Enum(SQLITE, POSTGRES)
 	DBName             = App.Flag("db-name", "name of database").Default(DEFAULT_DB_NAME).String()
@@ -177,6 +180,8 @@ const POSTGRES string = "postgres"
 const DEFAULT_RULES_DIR = "./rules"
 const DEFAULT_OUTPUT_DIR = "csa-reports"
 const DEFAULT_REPORT_DIR = "Reports"
+const DEFAULT_EXPORT_DIR = "Exports"
+const DEFAULT_EXPORT_FILE_NAME = "export"
 const DEFAULT_MODELS_DIR = "./scoring-models"
 const RULE_BOOTSTRAP_TEMPLATE = "BootstrapRulesTemplate.txt"
 const BIN_BOOTSTRAP_TEMPLATE = "BootstrapBinsTemplate.txt"
